Move PlayStation 2 specific logic to OpenKh.Ps2
