+++ OpenKh.Engine/Parsers/Kddf2/Kkdf2MdlxParser.cs
@@ -1,6 +1,7 @@
 using OpenKh.Common;
 using OpenKh.Engine.Maths;
 using OpenKh.Kh2;
+using OpenKh.Ps2;
 using System;
 using System.Collections.Generic;
 using System.IO;
