+++ OpenKh.Engine/Parsers/Kddf2/VU1Simulation.cs
@@ -1,4 +1,5 @@
 ﻿using OpenKh.Engine.Maths;
+using OpenKh.Ps2;
 using System;
 using System.IO;
 using System.Numerics;
