+++ OpenKh.sln
@@ -106,7 +106,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenKh.Tools.IdxImg", "Open
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenKh.Tests.Commands", "OpenKh.Tests.Commands\OpenKh.Tests.Commands.csproj", "{41706B1E-DFF9-472F-AD74-A936B10D80F6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenKh.Tests.Engine", "OpenKh.Tests.Engine\OpenKh.Tests.Engine.csproj", "{198657C3-DC06-4AA6-A594-6D4D82218CA8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenKh.Tests.Engine", "OpenKh.Tests.Engine\OpenKh.Tests.Engine.csproj", "{198657C3-DC06-4AA6-A594-6D4D82218CA8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenKh.Ps2", "OpenKh.Ps2\OpenKh.Ps2.csproj", "{8D0CBCCA-3F3E-4775-B001-2DDA4D962A8A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -823,6 +825,22 @@ Global
 		{198657C3-DC06-4AA6-A594-6D4D82218CA8}.Release|x64.Build.0 = Release|Any CPU
 		{198657C3-DC06-4AA6-A594-6D4D82218CA8}.Release|x86.ActiveCfg = Release|Any CPU
 		{198657C3-DC06-4AA6-A594-6D4D82218CA8}.Release|x86.Build.0 = Release|Any CPU
+		{8D0CBCCA-3F3E-4775-B001-2DDA4D962A8A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8D0CBCCA-3F3E-4775-B001-2DDA4D962A8A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8D0CBCCA-3F3E-4775-B001-2DDA4D962A8A}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{8D0CBCCA-3F3E-4775-B001-2DDA4D962A8A}.Debug|ARM.Build.0 = Debug|Any CPU
+		{8D0CBCCA-3F3E-4775-B001-2DDA4D962A8A}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{8D0CBCCA-3F3E-4775-B001-2DDA4D962A8A}.Debug|x64.Build.0 = Debug|Any CPU
+		{8D0CBCCA-3F3E-4775-B001-2DDA4D962A8A}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{8D0CBCCA-3F3E-4775-B001-2DDA4D962A8A}.Debug|x86.Build.0 = Debug|Any CPU
+		{8D0CBCCA-3F3E-4775-B001-2DDA4D962A8A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8D0CBCCA-3F3E-4775-B001-2DDA4D962A8A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8D0CBCCA-3F3E-4775-B001-2DDA4D962A8A}.Release|ARM.ActiveCfg = Release|Any CPU
+		{8D0CBCCA-3F3E-4775-B001-2DDA4D962A8A}.Release|ARM.Build.0 = Release|Any CPU
+		{8D0CBCCA-3F3E-4775-B001-2DDA4D962A8A}.Release|x64.ActiveCfg = Release|Any CPU
+		{8D0CBCCA-3F3E-4775-B001-2DDA4D962A8A}.Release|x64.Build.0 = Release|Any CPU
+		{8D0CBCCA-3F3E-4775-B001-2DDA4D962A8A}.Release|x86.ActiveCfg = Release|Any CPU
+		{8D0CBCCA-3F3E-4775-B001-2DDA4D962A8A}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -872,6 +890,7 @@ Global
 		{1B89F510-3B9A-4A3B-A08F-4FFB5C5B817B} = {402B2669-D594-4DFA-965B-02A92626ADC6}
 		{41706B1E-DFF9-472F-AD74-A936B10D80F6} = {9C52D787-9819-4B89-989B-EEA6FEB20731}
 		{198657C3-DC06-4AA6-A594-6D4D82218CA8} = {939E8104-3A37-41FC-A411-CAEE67A803B6}
+		{8D0CBCCA-3F3E-4775-B001-2DDA4D962A8A} = {0FB7CD6A-EE31-467D-A590-267DCD028618}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {1DF3960B-4FE5-4E56-92D4-2FC0A912E823}
